enh: Organizing condenser implementations (#6529)
