+++ openhands/memory/condenser/__init__.py
@@ -1,0 +1,3 @@
+from openhands.memory.condenser.condenser import Condenser, get_condensation_metadata
+
+__all__ = ['Condenser', 'get_condensation_metadata']
